게시글 등록시 체크여부로 letter, bbs테이블 분기 등록처리

+++ src/main/java/com/example/mujeans/repository/board/BoardRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BoardRepository extends JpaRepository<BoardDTO, Long> {
+public interface BoardRepository extends JpaRepository<BoardDTO, Integer> {
     List<BoardDTO> findAll();
     List<BoardDTO> findByUseYn(String useYn);
     BoardDTO save(BoardDTO boardDTO);
